Split 'SchemaField' out into its own module.
Reduces lines in 'test_table.py' to appease pylint's arbitrary ceiling.
