+++ docs/bigquery-schema.rst
@@ -1,0 +1,7 @@
+Schemas
+~~~~~~~
+
+.. automodule:: gcloud.bigquery.schema
+  :members:
+  :show-inheritance:
+
