+++ gcloud/bigquery/__init__.py
@@ -25,7 +25,7 @@ from gcloud.bigquery.client import Client
 from gcloud.bigquery.connection import Connection
 from gcloud.bigquery.dataset import AccessGrant
 from gcloud.bigquery.dataset import Dataset
-from gcloud.bigquery.table import SchemaField
+from gcloud.bigquery.schema import SchemaField
 from gcloud.bigquery.table import Table
